refactor(tui): split compat models
